更新构建脚本,修复已知问题

+++ install/build.py
@@ -16,7 +16,6 @@ PODMAN = 'podman'
 
 CONTAINER_CMD = None
 
-
 LOG = logging.getLogger(__name__)
 
 
@@ -45,7 +44,11 @@ def main():
     CONTAINER_CMD = shell.get_container_impl(impl=args.container_cmd)
 
     whl_file_path = os.path.abspath(args.whl_file)
-    LOG.info('whl file path: %s', whl_file_path)
+    whl_name = os.path.basename(whl_file_path)
+    requirements = 'requirements.txt'
+
+    LOG.info('=====> check whl file and version')
+    LOG.debug('whl file path: %s', whl_file_path)
     if not os.path.exists(whl_file_path):
         LOG.error('whl file %s is not exists', args.whl_file)
         return 1
@@ -53,7 +56,6 @@ def main():
         LOG.error('invalid path %s, it must be a file')
         return 1
 
-    whl_name = os.path.basename(whl_file_path)
     try:
         whl_version = whl_name.split('-')[1]
         LOG.debug('whl version is %s', whl_version)
@@ -61,31 +63,65 @@ def main():
         LOG.error(e)
         return 1
 
-    target = f'vstackboard:{whl_version}'
-    shutil.copy(whl_file_path, INSTALL_DIR)
-    requirements = 'requirements.txt'
+    LOG.info('=====> make fornedend')
+    frontend_path = os.path.join('forward', 'dist')
+    if not os.path.exists(frontend_path):
+        LOG.error('frontend project path %s is not exists', frontend_path)
+        return 1
 
+    frontend_path = os.path.join('forward', 'dist')
+    frontend_name = 'forward'
+
+    LOG.info('=====> prepare build resources')
+    forward_dst = os.path.join(INSTALL_DIR , frontend_name)
+    if os.path.exists(forward_dst):
+        shutil.rmtree(forward_dst)
+    shutil.copytree(frontend_path, forward_dst)
+    shutil.copy(whl_file_path, INSTALL_DIR)
     shutil.copy(requirements, INSTALL_DIR)
 
+    LOG.info('=====> update config.json backend_url')
+    config_json = os.path.join(INSTALL_DIR, frontend_name, 'config.json')
+    if not os.path.exists(config_json):
+        LOG.warning('config.json file %s not exists', config_json)
+    else:
+        import json
+        with open(config_json) as f:
+            content = json.load(f)
+        if 'backend_url' in content:
+            content['backend_url'] = ''
+        with open(config_json, 'w') as f:
+            json.dump(content, f, indent=4)
+
+    LOG.info('=====> build docker image')
     os.chdir(INSTALL_DIR)
-    LOG.info('change dir to %s', os.getcwd())
+    LOG.debug('changed dir to %s', os.getcwd())
 
     if not os.path.exists(whl_name):
         LOG.error('file %s not exists', whl_name)
         return 1
+
+    target = f'vstackboard:{whl_version}'
     try:
-        CONTAINER_CMD.build('./', network='host', target=target,
-                            no_cache=args.no_cache,
-                            build_args=[f'PACKAGE_NAME={whl_name}'])
+        CONTAINER_CMD.build(
+            './', network='host', target=target,
+            no_cache=args.no_cache,
+            build_args=[f'PACKAGE_NAME={whl_name}',
+                        f'FORWARD_PACKAGE_NANME={frontend_name}'])
+
         LOG.info('image build success')
     except Exception:
         LOG.exception('build failed!')
         return
     finally:
-        for file in [os.path.join(INSTALL_DIR, requirements), whl_name]:
+        for file in [os.path.join(INSTALL_DIR, requirements), whl_name,
+                     frontend_name]:
             if os.path.exists(file):
                 LOG.debug('clean file %s', file)
-                os.remove(file)
+                if os.path.isdir(file):
+                    shutil.rmtree(file)
+                else:
+                    os.remove(file)
 
     try:
         for registry in args.push_registry or []:
